Create sql6-4.sql
On the Course table there is an old course that is no longer lectured and that has now been replaced with a newer one. Delete this old course from the table.

The details of the record to be removed are given below. NOTE: There is also another course with the same name now on the table--it is the newer course that replaces the to-be-deleted one. Make sure you do not remove this newer course while deleting the older one.

The old record to be deleted:

Attribute     Value

courseid       1010  (primary key value)
Name           Introduction to databases
Starts         2nd September, 2006
Lecturer       No lecturer