Create sql8-3.sql
Provide a query that prints the prices and titles (in that order) for all books, but so that the books are ordered by their price such that the most expensive book is the first and the cheapest one is the last (i.e. decreasing order).
IN ADDITION TO THIS omit all books that cost less than 30.