Introduction to SQL ex1
In this very first SQL exercise, you'll see how the SQL environment works. Your task is to write the correct answer and submit it, and if it is correct, your answer is accepted by the system and the exercise is marked as done, and you'll be automatically taken to the next exercise.
The solution for this task is provided for you below, and you only need to write it in the text field (you can test the editor that way, too) and submit the answer. Should there be any problems with your submission, you'll receive an error report and be taken back so you can fix your code and try again. You can test this by issuing a wrong answer on purpose.
Task description:
This exercise is related to the Books database, description of which is found in the references
Write an SQL query that fetches the title, price and the number of pages (in that order) for each book published prior to 1995. The books are to be ordered alphabetically by their names.